Reorganizing code
This is as per the recommendation from issue#60.
