+++ Swift/Sources/Astronomy/ShadowLength.swift
@@ -1,0 +1,13 @@
+//
+//  ShadowLength.swift
+//  Adhan
+//
+//  Copyright © 2018 Batoul Apps. All rights reserved.
+//
+
+import Foundation
+
+enum ShadowLength: Double {
+    case single = 1.0
+    case double = 2.0
+}
