+++ Swift/Sources/CalculationParameters.swift
@@ -1,0 +1,50 @@
+//
+//  CalculationParameters.swift
+//  Adhan
+//
+//  Copyright © 2018 Batoul Apps. All rights reserved.
+//
+
+import Foundation
+
+/* All customizable parameters for calculating prayer times */
+public struct CalculationParameters {
+    public var method: CalculationMethod = .other
+    public var fajrAngle: Double
+    public var ishaAngle: Double
+    public var ishaInterval: Int = 0
+    public var madhab: Madhab = .shafi
+    public var highLatitudeRule: HighLatitudeRule = .middleOfTheNight
+    public var adjustments: PrayerAdjustments = PrayerAdjustments()
+
+    init(fajrAngle: Double, ishaAngle: Double) {
+        self.fajrAngle = fajrAngle
+        self.ishaAngle = ishaAngle
+    }
+
+    init(fajrAngle: Double, ishaInterval: Int) {
+        self.init(fajrAngle: fajrAngle, ishaAngle: 0)
+        self.ishaInterval = ishaInterval
+    }
+
+    init(fajrAngle: Double, ishaAngle: Double, method: CalculationMethod) {
+        self.init(fajrAngle: fajrAngle, ishaAngle: ishaAngle)
+        self.method = method
+    }
+
+    init(fajrAngle: Double, ishaInterval: Int, method: CalculationMethod) {
+        self.init(fajrAngle: fajrAngle, ishaInterval: ishaInterval)
+        self.method = method
+    }
+
+    func nightPortions() -> (fajr: Double, isha: Double) {
+        switch self.highLatitudeRule {
+        case .middleOfTheNight:
+            return (1/2, 1/2)
+        case .seventhOfTheNight:
+            return (1/7, 1/7)
+        case .twilightAngle:
+            return (self.fajrAngle / 60, self.ishaAngle / 60)
+        }
+    }
+}
