+++ Swift/Sources/Enums/Prayer.swift
@@ -1,0 +1,18 @@
+//
+//  Prayer.swift
+//  Adhan
+//
+//  Copyright © 2018 Batoul Apps. All rights reserved.
+//
+
+import Foundation
+
+public enum Prayer {
+    case fajr
+    case sunrise
+    case dhuhr
+    case asr
+    case maghrib
+    case isha
+    case none
+}
