+++ Swift/Sources/Extensions/Calendar+Extension.swift
@@ -1,0 +1,18 @@
+//
+//  Calendar+Extension.swift
+//  Adhan
+//
+//  Copyright © 2018 Batoul Apps. All rights reserved.
+//
+
+import Foundation
+
+extension Calendar {
+
+    /// All calculations are done using a gregorian calendar with the UTC timezone
+    static let gregorianUTC: Calendar = {
+        var cal = Calendar(identifier: .gregorian)
+        cal.timeZone = TimeZone(identifier: "UTC")!
+        return cal
+    }()
+}
