+++ Swift/Sources/PrayerAdjustment.swift
@@ -1,0 +1,27 @@
+//
+//  PrayerAdjustment.swift
+//  Adhan
+//
+//  Copyright © 2018 Batoul Apps. All rights reserved.
+//
+
+import Foundation
+
+/* Adjustment value for prayer times, in minutes */
+public struct PrayerAdjustments {
+    public var fajr: Int
+    public var sunrise: Int
+    public var dhuhr: Int
+    public var asr: Int
+    public var maghrib: Int
+    public var isha: Int
+
+    public init(fajr: Int = 0, sunrise: Int = 0, dhuhr: Int = 0, asr: Int = 0, maghrib: Int = 0, isha: Int = 0) {
+        self.fajr = fajr
+        self.sunrise = sunrise
+        self.dhuhr = dhuhr
+        self.asr = asr
+        self.maghrib = maghrib
+        self.isha = isha
+    }
+}
